+++ lib/Error.php
@@ -1,5 +1,7 @@
 <?php
-class Laposta_Error extends Exception {
+namespace Laposta;
+
+class Error extends Exception {
 
 	public function __construct($message=null, $http_status=null, $http_body=null, $json_body=null) {
