+++ lib/Member.php
@@ -1,5 +1,7 @@
 <?php
-class Laposta_Field extends Laposta_Resource {
+namespace Laposta;
+
+class Member extends Resource {
 
 	private $list_id;
 
@@ -10,10 +12,10 @@ class Laposta_Field extends Laposta_Resource {
 		parent::__construct(get_class());
 	}
 
-	public function get($field_id) {
+	public function get($member_id) {
 
 		return parent::connect(array(
-			'path' => array($field_id),
+			'path' => array($member_id),
 			'parameters' => array('list_id' => $this->list_id)
 			)
 		);
@@ -30,22 +32,22 @@ class Laposta_Field extends Laposta_Resource {
 		);
 	}
 
-	public function update($field_id, $data) {
+	public function update($member_id, $data) {
 
 		// add list_id to data
 		$data['list_id'] = $this->list_id;
 
 		return parent::connect(array(
-			'path' => array($field_id),
+			'path' => array($member_id),
 			'post' => $data
 			)
 		);
 	}
 
-	public function delete($field_id) {
+	public function delete($member_id) {
 
 		return parent::connect(array(
-			'path' => array($field_id),
+			'path' => array($member_id),
 			'parameters' => array('list_id' => $this->list_id),
 			'method' => 'DELETE'
 			)
